Remove coloredtestrunner for easier to read verbose test output
