+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix incompatibility issue with downstream github-release tool used to deploy releases to GitHub. @rly (#1245)
 - Fix issue with Sphinx gallery. @rly
 - Add citation information to documentation and support for duecredit tool. @rly
+- Remove use of ColoredTestRunner for more readable verbose test output. @rly
 - Add support for nwb-schema 2.3.0. @rly (#1245, #1330)
   - Add optional `waveforms` column to the `Units` table.
   - Add optional `strain` field to `Subject`.
